PDC: openmp: dot-product: Convert to `long long int`

+++ ParallelDistributedComputing/openmp/dot-product.c
@@ -16,20 +16,20 @@ int main(int argc, char *argv[])
     printf(GREEN "[-] Enter value of n: " RESET);
     scanf("%d", &n);
 
-    int *a = (int *)malloc(n * sizeof(int));
-    int *b = (int *)malloc(n * sizeof(int));
+    long long int *a = (long long int *)malloc(n * sizeof(long long int));
+    long long int *b = (long long int *)malloc(n * sizeof(long long int));
 
     printf(YELLOW "[+] Enter elements of vector a: " RESET);
-    for (int i = 0; i < n; i++) scanf("%d", &a[i]);
+    for (int i = 0; i < n; i++) scanf("%lld", &a[i]);
     
     printf(YELLOW "[+] Enter elements of vector b: " RESET);
-    for (int i = 0; i < n; i++) scanf("%d", &b[i]);
+    for (int i = 0; i < n; i++) scanf("%lld", &b[i]);
 
-    int dot_product = 0;
+    long long int dot_product = 0;
     #pragma omp parallel for reduction(+:dot_product)
     for (int i = 0; i < n; i++) dot_product += a[i] * b[i];
 
-    printf(BLUE "[✓] Dot product of vectors a and b is: %d\n" RESET, dot_product);
+    printf(BLUE "[✓] Dot product of vectors a and b is: %lld\n" RESET, dot_product);
 
     free(a);
     free(b);
